Refactor: move extract schema meta inside Schema constructor
