+++ src/helpers/index.ts
@@ -1,9 +1,11 @@
 import queryHelpers from './queryhelpers';
 import validation from './validation';
 import populateHelpers from './populateHelpers';
+import schemaHelpers from './schema.helpers';
 
 export default {
   queryHelpers,
   validation,
   populateHelpers,
+  schemaHelpers,
 };
